% Generated by roxygen2 (4.0.2): do not edit by hand
\name{address_lookup}
\alias{address_lookup}
\title{Reverse Geocoding: Adresse from Longitude and Latitude}
\usage{
address_lookup()
}
\value{
an object of class UrlData
}
\description{
The UrlData object provides a resource 'AddressLookup' which takes two parameters
\code{lat} and \code{lon} and returns an approximate adress for this coordinates.
}
\details{
Based on Stackoverflow solution by Jochem Donkers.
}
\references{
\href{http://stackoverflow.com/questions/13905098/how-to-get-the-longitude-and-latitude-coordinates-from-a-city-name-and-country-i}{Stackoverflow}
}

