% Generated by roxygen2 (4.0.2): do not edit by hand
\name{allensbach}
\alias{allensbach}
\title{Election polls for Germany}
\usage{
allensbach(resource = "allensbach")
}
\arguments{
\item{resource}{the resource name (character)}
}
\description{
This function exposes an interface to the
IfD/Allensbach website's poll data.
}
\details{
It serves as a proof of concept for the \code{urldata} function.
}
\seealso{
\code{\link{urldata}}
}

