\docType{methods}
\name{as.character}
\alias{as.character}
\alias{as.character,DirectoryLocation-method}
\alias{as.character,Resource-method}
\title{as.character method for Resource}
\arguments{
  \item{a}{Resource object}

  \item{a}{Directory object}
}
\description{
  This method returns the class of itself. It is just the
  identifier of the resource.

  This method returns the path to the directory it
  represents.
}

