\docType{data}
\name{be_bathing_areas}
\alias{be_bathing_areas}
\title{be_bathing_areas -- geocoordinates of bathing areas around Berlin}
\usage{
  be_bathing_areas()

  be_bathing_areas
}
\description{
  be_bathing_areas -- geocoordinates of bathing areas
  around Berlin
}
\references{
  \url{http://www.berlin.de/badegewaesser/baden-details/index.php/index/all.kml?q=}
  \url{http://code.google.com/intl/de-DE/apis/kml/documentation/kmlreference.html}
}
\keyword{datasets}

