\docType{data}
\name{be_bathing_water_quality}
\alias{be_bathing_water_quality}
\title{Water quality at bathing spots in Berlin}
\usage{
  be_bathing_water_quality()

  be_bathing_water_quality
}
\description{
  E.coli pro 100 ml, Intestinale Enterokokken pro 100 ml,
  Sichttiefe in cm
}
\references{
  \url{http://www.berlin.de/badegewaesser/baden-details/index.php/index/all.json?q=}
  \url{http://support.berlin.de/wiki/index.php/SimpleSearch_JSON}
}
\keyword{datasets}

