\name{cityCoords}
\alias{cityCoords}
\title{Longitude and Latitude for Cities}
\usage{
  cityCoords(country = "DE")
}
\arguments{
  \item{country}{two-character country code, default 'DE'}
}
\value{
  an object of class UrlData
}
\description{
  Based on Stackoverflow solution by Jochem Donkers.
}
\references{
  \url{http://stackoverflow.com/questions/13905098/how-to-get-the-longitude-and-latitude-coordinates-from-a-city-name-and-country-i}
}

