% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{CsvData-class}
\alias{CsvData-class}
\title{CsvData -- tabular data and its metadata}
\description{
This class provides the infrastructure to
read and write tabular data, including
metadata on the dataset as a whole and
metadata on the columns of the dataset.
}
\details{
Metadata on the dataset that are not interpreted are
title, description, license, publisher, keywords.
Metadata on the dataset that is interpreted are
update.scheme (always, never, and everything that
can be passed to seq.POSIXt as by parameter),
update.method (lastmod, fullreplace),
encoding (an element of iconvlist()), type (csv, csv2),
modified (automatically set, YYYY-mm-dd HH:MM:SS UTC format).

Metadata on a dataset column that are currently interpreted are name,
type (logical, integer, numeric, complex, character, raw, factor, Date, POSIXct),
format (for Date and POSIXct, format string for strptime).
}
\examples{
getSlots("CsvData")
}
\seealso{
\code{\link{csvdata}}, \code{\link{read.csvdata}}
}

