\name{eurostat_data}
\alias{eurostat_data}
\alias{eurostat_dicts}
\alias{eurostat_toc}
\title{eurostat_data -- creates UrlData object for accessing Eurostat datasets.}
\usage{
  eurostat_data()

  eurostat_toc()

  eurostat_dicts()
}
\description{
  eurostat_data -- creates UrlData object for accessing
  Eurostat datasets.

  eurostat_toc -- creates UrlData object for accessing
  Eurostat "table of contents", i.e. a description of
  datasets.

  eurostat_dicts -- creates UrlData object for accessing
  Eurostat "dictionaries", i.e. a description of
  indicators.
}

