\name{DataStoreDriver-class}
\alias{DataStoreDriver-class}
\title{DataStoreDriver -- A wrapper to SQLiteDriver}
\description{
  In order to define a dbConnect method for
  SQLiteDataStore, a custom class derived from SQLiteDriver
  is needed.
}
\details{
  This class is for internal purposes only.
}

