\name{Dbpedia-class}
\alias{Dbpedia-class}
\title{A class for querying Dbpedia.org}
\description{
  This class defines some resources at dbpedia. See
  \code{queries(dbpedia())} for a list of resources.
}
\examples{
\dontrun{
  dbp <- dbpedia()
  queries(dbpedia)
  query(dbp, "Nuts1")
}
}

