\name{dbpedia}
\alias{dbpedia}
\title{A data object for querying dbpedia}
\usage{
  dbpedia(lang = "")
}
\arguments{
  \item{lang}{two-character language code for the dbpedia,
  default ''}
}
\value{
  a Dbpedia object, inherited from Xsparql
}
\description{
  A data object for querying dbpedia
}

