% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{dependencies}
\alias{dependencies}
\alias{dependencies,ResFunc-method}
\alias{dependencies,Xdata-method}
\title{Dependencies of a Xdata object}
\usage{
dependencies(self)

\S4method{dependencies}{Xdata}(self)

\S4method{dependencies}{ResFunc}(self)
}
\arguments{
\item{self}{an Xdata object}
}
\description{
The \code{dependencies} method returns a list of character or
list elements that define the resources the object depends on.
}
\details{
By default, NULL is returned.

Inherited classes should override this method
if necessary.
}

