% Generated by roxygen2 (4.0.2): do not edit by hand
\name{enipedia}
\alias{enipedia}
\title{Interface to the Enipedia Project}
\usage{
enipedia()
}
\value{
a Mashup object
}
\description{
Enipedia is an active exploration into the applications of wikis and the semantic
web for energy and industry issues. Through this we seek to create a collaborative
environment for discussion, while also providing the tools that allow for data from
different sources to be connected, queried, and visualized from different perspectives.
}
\details{
This function creates an datamart with selected queries to the SPARQL endpoint of
the enipedia project.
}
\seealso{
\code{\link{enipedia}}, \code{\link{xsparql}}
}

