% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{enipedia_countries}
\alias{enipedia_countries}
\title{Translation map between Enipedia country codes and ISO Codes}
\format{CSV file (western european style)}
\source{
\href{http://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{Wikipedia for ISO country codes},
\href{http://enipedia.tudelft.nl/wiki/Using_SPARQL_with_Enipedia}{Enipedia for Enipedia names}
}
\usage{
enipedia_countries
}
\description{
This dataset maps the 2-digit ISO 3166-1 alpha-2 country code to the country name used by enipedia.tudelf.nl
}

