\docType{data}
\name{eurostat}
\alias{eurostat}
\title{Eurostat data object.}
\usage{
  eurostat()

  eurostat
}
\value{
  an object of class Mashup
}
\description{
  From Eurostat's webpage (see references): You can
  download individual datasets or the complete database by
  using the bulk download facility. On the bulk download
  you will find: - all information updated twice a day, at
  11:00 and 23:00, (at this time the service is usually not
  available, K.W.) - the datasets in tsv (tab separated
  values) (this is what is used by this R code, K.W.), dft
  and sdmx format, which can be easily used to import the
  data in a tool of your choice, - a manual containing all
  detailed information on the bulkdownload facility, - the
  table of contents that includes the list of the datasets
  available, - the "dictionaries" of all the coding systems
  used in the datasets.
}
\references{
  \url{http://epp.eurostat.ec.europa.eu/}
}
\keyword{datasets}

