\docType{data}
\name{evs2008.lvl2-data}
\alias{evs2008.lvl2}
\alias{evs2008.lvl2-data}
\title{German Income and Expenditure Survey 2008 on private spendings, differentiated by household type and household income.}
\description{
  German Income and Expenditure Survey 2008 on private
  spendings, differentiated by household type and household
  income.
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}
\references{
  Statistisches Bundesamt: Wirtschaftsrechnungen.
  Einkommens- und Verbrauchsstichprobe. Einnahmen und
  Ausgaben privater Haushalte. Fachserie 15 Heft 4.
}
\keyword{datasets}

