% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{evs2008.lvl2}
\alias{evs2008.lvl2}
\title{German Income and Expenditure Survey 2008 on private spendings, differentiated by household type and household income.}
\format{CSV file (western european style)}
\source{
Statistisches Bundesamt: Wirtschaftsrechnungen. Einkommens- und Verbrauchsstichprobe. Einnahmen und Ausgaben privater Haushalte. Fachserie 15 Heft 4.
}
\usage{
evs2008.lvl2
}
\description{
Germany's Sample survey of income and expenditure (Einkommens- und Verbrauchsstichprobe, EVS) is conducted by the
Federal Statistical Office. The data provided here is processed and is not identical with der Federal Office' data.
Some information is lost by the processing. If you want more and/or more accurate data,
contact the Federal Statistical Office.
}

