\name{expenditures}
\alias{expenditures}
\title{Constructor for Evs data source class}
\usage{
  expenditures()
}
\value{
  an object of class "Evs"
}
\description{
  Germany's Sample survey of income and expenditure
  (Einkommens- und Verbrauchsstichprobe, EVS) is conducted
  by the Federal Statistical Office. The data provided here
  is processed and is not identical with der Federal
  Office' data. Some information is lost by the processing.
  If you want more and/or more accurate data, contact the
  Federal Statistical Office.
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

