% Generated by roxygen2 (4.0.2): do not edit by hand
\name{google.oauth2}
\alias{google.oauth2}
\title{Google OAuth2 initial authentication}
\usage{
google.oauth2(scope, client_id = getOption("datamart.client_id"),
  client_secret = getOption("datamart.client_secret"), name = "",
  clss = "GoogleOAuth2", curl = RCurl::getCurlHandle(), verbose = TRUE)
}
\arguments{
\item{scope}{either shortname ("blogger") or URL of the scope}

\item{client_id}{client id. Defaults to getOption("datamart.client_id")}

\item{client_secret}{client secret. Defaults to getOption("datamart.client_secret")}

\item{name}{currently ignored}

\item{clss}{the class to create. Must be derived from GoogleOAuth2}

\item{curl}{curl.handle object}

\item{verbose}{prints diagnostic messages on the way}
}
\value{
GoogleOAuth2 object
}
\description{
This function redirects the user to a consent screen where the user can
grant or revoke access for the provided scope.
}
\seealso{
\code{\link{read.google.oauth2}}, \code{\link{google.oauth2}}
}

