% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{GoogleOAuth2-class}
\alias{GoogleOAuth2-class}
\title{A class for managing Google OAuth2 workflow for installed apps}
\description{
This class is aimed to manage the process where the user
grants Google account access to your R application.
}
\details{
There are two basic ways to create an GoogleOAuth2 object.
First, you can use \code{google.oauth2} to initially register
your application with google. You may store this authentication
information using \code{put}. Second, you can use \code{read.google.oauth2}
to load authentication information from a previously written file.
}
\references{
\href{https://developers.google.com/accounts/docs/OAuth2InstalledApp?hl=de}{Google}
}
\seealso{
\code{\link{google.oauth2}},
}

