\name{Gscholar-class}
\alias{Gscholar-class}
\title{Gscholar -- A class for querying Google Scholar}
\description{
  This class uses Google Scholar to determine how many
  publications contain a given search term in a given time
  span (resolution: year). The idea for the class comes
  from a blog post on librestats, see references. This
  class is likely to change in future versions.
}
\references{
  \url{http://librestats.com/2012/04/12/statistical-software-popularity-on-google-scholar/}
}

