\name{internalData}
\alias{internalData}
\title{Constructor for InternalData objects}
\usage{
  internalData(name, package = NULL, clss = "InternalData")
}
\arguments{
  \item{name}{name of the dataset. Required.}

  \item{package}{name of the package where the dataset is
  located. Default NULL.}

  \item{clss}{name of the class to create. Default
  InternalData, must be inherited from this class.}
}
\description{
  Constructor for InternalData objects
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

