\docType{data}
\name{iwv_online}
\alias{iwv_online}
\title{website traffic from IWV}
\usage{
  iwv_online()

  iwv_online
}
\value{
  UrlData3 object
}
\description{
  website traffic as tracked by iwv online. Use yyyymm for
  query() as resource.
}
\references{
  \url{http://en.wikipedia.org/wiki/Informationsgemeinschaft_zur_Feststellung_der_Verbreitung_von_Werbetraegern}
}
\keyword{datasets}

