\docType{data}
\name{mauna_loa}
\alias{mauna_loa}
\title{CO2 data from Mauna Loa}
\usage{
  mauna_loa()

  mauna_loa
}
\value{
  UrlData
}
\description{
  Access to monthly and annual mean CO2 concentration.
}
\references{
  \url{http://www.esrl.noaa.gov/gmd/ccgg/trends/co2_data_mlo.html}
  \url{http://tamino.wordpress.com/2012/06/07/co2-sanity/}
}
\keyword{datasets}

