\name{mdreport}
\alias{mdreport}
\title{Constructor for MdReport objects}
\usage{
  mdreport(tpl, name = "", clss = "MdReport",
    verbose = getOption("verbose"), ...)
}
\arguments{
  \item{tpl}{path to markdown template file}

  \item{name}{name of the Report, default ''}

  \item{verbose}{diagnostic messages T/F}

  \item{clss}{class of the object, default 'MdReport'}

  \item{...}{number of targets}
}
\value{
  generic
}
\description{
  see class MdReport for details.
}

