% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MemoryLocation-class}
\alias{MemoryLocation-class}
\title{S4 base class to represent output in Memory}
\description{
The \code{MemoryLocation} class represents the place in the memory of
the current R process.
}
\examples{
getSlots("MemoryLocation")
}

