% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{MemoryLocation2-class}
\alias{MemoryLocation2-class}
\alias{memloc}
\title{S4 base class to represent output in Memory}
\usage{
memloc(clss = "MemoryLocation2", ...)
}
\arguments{
\item{clss}{class to construct. Defaults to MemoryLocation2.}

\item{...}{initial objects that are contained in the location}
}
\description{
The \code{MemoryLocation2} class represents the place in the memory of
the current R process.

Constructor for MemoryLocation2 objects
}
\examples{
getSlots("MemoryLocation2")
}
\seealso{
\code{\link{memloc}}
}

