\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,Blogger-method}
\alias{meta,DirectoryLocation-method}
\alias{meta,Pastebin-method}
\alias{meta,Xdata-method}
\title{Verbose list of resources}
\arguments{
  \item{self}{an Location object}
}
\description{
  The \code{meta} method returns a data.frame with meta
  information entities available at the location. By
  default, zero rows are returned.
}
\details{
  Inherited classes should override this method if
  necessary.
}

