\name{netConnectGermany}
\alias{netConnectGermany}
\title{Basic Price Information on natural Gas in Germany}
\usage{
  netConnectGermany(clss = "UrlData3")
}
\arguments{
  \item{clss}{Class name for the object, default UrlData2}
}
\value{
  UrlData2
}
\description{
  Basic Price Information on natural Gas in Germany
}
\references{
  \url{http://datenservice.net-connect-germany.de/Dokumente/NCG_XML_Interface_V1.6_de.pdf}
}

