% Generated by roxygen2 (4.0.2): do not edit by hand
\name{openei}
\alias{openei}
\title{Interface to the OpenEI platform}
\usage{
openei()
}
\description{
OpenEI is growing into a global leader in the energy data realm - specifically
analyses on renewable energy and energy efficiency. The platform is a wiki,
similar to Wikipedia's Wiki, and offers an SPARQL endpoint.
}
\details{
The \code{openei} function provides an datamart with predefined queries on
this SPARQL endpoint.
}
\references{
\href{http://chrisdavis.weblog.tudelft.nl/2010/06/12/data-mining-the-us-department-of-energy/}{Blogpost by Chris Davis},
\href{http://en.openei.org/wiki/OpenEI:About}{OpenEI website}
}
\seealso{
\code{\link{openei}}, \code{\link{xsparql}}
}

