\name{pkg}
\alias{pkg}
\title{Constructor for PkgTarget objects}
\usage{
  pkg(name, clss = "PkgTarget",
    verbose = getOption("verbose"), ...)
}
\arguments{
  \item{name}{name of the Report, default ''}

  \item{verbose}{diagnostic messages T/F}

  \item{clss}{class name, default 'PkgTarget'}

  \item{...}{number of targets}
}
\value{
  generic
}
\description{
  see class PkgTarget for details.
}

