\docType{methods}
\name{put}
\alias{put}
\alias{put,BlogPostTarget,Blogger-method}
\alias{put,FileTarget,DirectoryLocation-method}
\alias{put,MdFigure,DirectoryLocation-method}
\alias{put,MdFigure,MemoryLocation-method}
\alias{put,MdReport,Location-method}
\alias{put,SweaveReport,DirectoryLocation-method}
\alias{put,SweaveReport,missing-method}
\alias{put,Target,character-method}
\title{Put a target}
\arguments{
  \item{target}{an object of class Target or derived}

  \item{where}{an object of class Location or derived}
}
\description{
  This generic creates the target at the given location.
}

