% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.csvdata}
\alias{read.csvdata}
\title{Read CSV from local directory or Internet}
\usage{
read.csvdata(resource, location = NULL, name = resource,
  update.fct = function(csv) return(data.frame()), verbose = TRUE,
  clss = "CsvData")
}
\arguments{
\item{resource}{the name of the resource. Required.}

\item{location}{either a Location object, or a character pointing to a local directory or an url. See details.}

\item{name}{the physical name of the resource. Defaults to resource.}

\item{update.fct}{function for updating the data. the update method and update interval is specified as meta data.
Default is a function that returns an empty data.frame.}

\item{verbose}{Print diagnostic messages, default is TRUE.}

\item{clss}{class to construct. Defaults to CsvData.}
}
\description{
It is lazy, i.e. it does not read the csv in until requested. It does read in the metadata, though.
If according to the dataset meta data an update is due, an update process is performed.
}

