% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.google.oauth2}
\alias{read.google.oauth2}
\title{Refreshed Google OAuth2 authentication}
\usage{
read.google.oauth2(fn)
}
\arguments{
\item{fn}{data filename}
}
\value{
GoogleOAuth2 object
}
\description{
This function reads a file previously created via \code{put(oauth2, filename)}.
}
\seealso{
\code{\link{google.oauth2}}
}

