% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resalias}
\alias{resalias}
\title{Phony ResFunc Objects}
\usage{
resalias(resource, alias_for, clss = "ResFunc")
}
\arguments{
\item{resource}{the name of the resource. Required.}

\item{alias_for}{name of the resource this object is a proxy for.}

\item{clss}{name of the class to create. Default ResFunc, must be inherited from this class.}
}
\description{
This function creates a ResFunc object that returns an
already existing resource.
}
\seealso{
\code{\link{resfunc}}
}

