% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{ResFunc-class}
\alias{ResFunc-class}
\title{ResFunc -- A class representing a "calculated" resource}
\description{
While the \code{UrlData} and \code{InternalData} classes
are examples for accessing actual data, the \code{ResFunc}
class represents resources that are not
physical data but for example simulated data or
data derived from physical data.
}
\details{
This class is intended to be used with the
\code{Mashup} class.
}
\seealso{
\code{\link{resfunc}}, \code{\link{resalias}}, \code{\link{datamart}}
}

