% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resfunc}
\alias{resfunc}
\title{Constructor for ResFunc objects}
\usage{
resfunc(resource, fun, depends = list(), clss = "ResFunc", ...)
}
\arguments{
\item{resource}{the name of the resource. Required.}

\item{fun}{a function that matches the signature function(self, resource, ...)}

\item{depends}{the names of the resources this function depends on.}

\item{clss}{name of the class to create. Default ResFunc, must be inherited from this class.}

\item{...}{additional parameters passed to the function when the resource is queried.}
}
\description{
This function creates an ResFunc object. When queried,
it returns the result of an function call.
}

