\docType{methods}
\name{show}
\alias{show}
\alias{show,Blogger-method}
\alias{show,DirectoryLocation-method}
\alias{show,FileTarget-method}
\alias{show,Pastebin-method}
\alias{show,Target-method}
\alias{show,Xdata-method}
\title{Show Method for Xdata classes}
\description{
  The \code{show} method for the Xdata class has been
  adapted to display the class name. Some inherited classes
  such DirectoryLocation or Blogger override this default
  definition.
}

