\name{SQLiteDataStore-class}
\alias{SQLiteDataStore-class}
\title{SQLiteDataStore -- A wrapper to SQLiteConnection}
\description{
  SQLiteDataStore provides different \code{dbReadTable} and
  \code{dbWriteTable} than its ancestor. Further methods
  for other data types than data.frame may be provided.
}
\details{
  Usually, objects of this class are created by
  \code{datastore}.
}

