\docType{data}
\name{statsoft_popularity}
\alias{statsoft_popularity}
\title{Statistical Software Popularity}
\usage{
  statsoft_popularity()

  statsoft_popularity
}
\value{
  a Gscholar object
}
\description{
  Some queries on Google Scholar on statistcal software,
  based on a blog post on librestats.com by Robert A.
  Muenchen.
}
\references{
  \url{http://librestats.com/2012/04/12/statistical-software-popularity-on-google-scholar/}
}
\keyword{datasets}

