% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strdecrypt}
\alias{strdecrypt}
\title{Obfuscate string}
\usage{
strdecrypt(message)
}
\arguments{
\item{message}{character or character vector}
}
\value{
character
}
\description{
reverts the action of strencrypt (not safe across machines!)
}

