% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strdehtml}
\alias{strdehtml}
\title{Decode HTML entities}
\usage{
strdehtml(s)
}
\arguments{
\item{s}{character or character vector}
}
\value{
character or character vector
}
\description{
Removes HTML named entities like &amp; and numbered entities like &#39; by replacing it with the corresponding character.
}

