% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strparse}
\alias{strparse}
\title{Parse named patterns}
\usage{
strparse(pat, x)
}
\arguments{
\item{pat}{named pattern}

\item{x}{character or character vector}
}
\value{
named character vector or matrix
}
\description{
code based on examples for regexpr()
}

