% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strrecode}
\alias{strrecode}
\title{Pattern-based recoding}
\usage{
strrecode(pats, repls, x, ...)
}
\arguments{
\item{pats}{vector of patterns}

\item{repls}{vector of replacements}

\item{x}{character or character vector}

\item{...}{additional parameter, passed to grepl}
}
\value{
replaced vector
}
\description{
Pattern-based recoding
}

