% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strtail}
\alias{strtail}
\title{Get the last n letters}
\usage{
strtail(s, n = 1)
}
\arguments{
\item{s}{character or character vector}

\item{n}{numeric, default 1}
}
\value{
character
}
\description{
if n>0, return the last n letters of x
if n<0, return all but the first abs(n) letters of x
}

