\name{SweaveReport-class}
\alias{SweaveReport-class}
\title{Wrapper for Sweave and pdf}
\description{
  The main S4 class in this framework is
  \code{SweaveReport}. You can create a report with
  \code{swvreport}, which takes a sweave file name, The
  generic method \code{build} can then be used to actually
  produce the report in pdf format.
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

