\name{texttarget}
\alias{texttarget}
\title{Constructor for TextReport objects}
\usage{
  texttarget(name = "", text.fct = "query", text.args,
    clss = "TextTarget")
}
\arguments{
  \item{name}{name of the Report, default ''}

  \item{text.fct}{function name}

  \item{text.args}{list of arguments past to text.fct}

  \item{clss}{class name, default 'TextTarget'}
}
\value{
  generic
}
\description{
  see class TextReport for details.
}

