\docType{data}
\name{tourdefrance}
\alias{tourdefrance}
\title{Tour de France data}
\usage{
  tourdefrance()

  tourdefrance
}
\value{
  an object of class UrlData
}
\description{
  Dataset collected by Martin Theusrus.
}
\references{
  \url{http://www.theusrus.de/blog/category/sports/tour-de-france/}
}
\keyword{datasets}

