\name{Twttr-class}
\alias{Twttr-class}
\title{A class for querying twitter}
\description{
  This class allows to query user timelines. Other
  resources have not been implemented yet.
}
\examples{
\dontrun{
  tw <- twttr(user="wagezudenken")
  query(tw, "User_timeline")
}
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

