\name{twttr}
\alias{twttr}
\title{Constructor for Twttr objects}
\usage{
  twttr(user = "")
}
\arguments{
  \item{user}{character, twitter screenname}
}
\description{
  Use twttr() to initialize. If no user parameter are
  provided, then the global options pft.user is used.
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

