% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uconvlist}
\alias{uconvlist}
\title{List unitsets and their units}
\usage{
uconvlist()
}
\value{
named list, names=Unitsets, values=Units in these Unitsets
}
\description{
The function lists the currently available
unitsets and the units supported by them.
}

