\docType{data}
\name{usets-data}
\alias{CrudeOil}
\alias{Energy}
\alias{GaseousFuel}
\alias{Length}
\alias{Mass}
\alias{Temperature}
\alias{usets}
\alias{usets-data}
\title{Measurement unit constants}
\description{
  Unit set for converting Length, Energy, Mass,
  GaseousFuel, Temperature, CrudeOil
}
\details{
  Usually you use this dataset indirect with uconv.
}
\references{
  \url{http://en.wikipedia.org/wiki/Mass}

  \url{http://en.wikipedia.org/wiki/Unit_of_length}

  Erdmann/Zweifel (2005): Energieoekonomik, Tab. 2.2, based
  on IEA conversion factors

  Erdmann/Zweifel (2005): Energieoekonomik, Tab. 9.1

  http://de.wikipedia.org/wiki/Temperatur

  Erdmann/Zweifel (2005): Energieoekonomik, Tab. 8.1
}
\keyword{datasets}

