\name{viz}
\alias{viz}
\title{Helper function for buildung figures}
\usage{
  viz(resource, ...)
}
\arguments{
  \item{resource}{character, name of the package}

  \item{...}{do not check for code/documentation
  mismatches}
}
\value{
  call
}
\description{
  A wrapper to querying a resource see class FigureTarget
}

