% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{WebLocation-class}
\alias{WebLocation-class}
\alias{webloc}
\title{Web location}
\usage{
webloc(path, clss = "WebLocation")
}
\arguments{
\item{path}{character, pointing to an existing web directory. Required.}

\item{clss}{character, optional class name. Default is "WebLocation".}
}
\description{
Read-only web folder

This function returns a WebLocation object that can be used as a read-only folder.
}
\details{
The queries and meta methods do not work for this class.
}
\examples{
getSlots("WebLocation")
}
\seealso{
\code{\link{webloc}}
}

