\docType{data}
\name{wikisites}
\alias{wikisites}
\title{List of the biggest Mediawiki sites}
\usage{
  wikisites()

  wikisites
}
\value{
  data.frame
}
\description{
  This function is like to disappear in future versions.
}
\references{
  \url{http://s23.org/wikistats/}
}
\keyword{datasets}

