% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Xdata-class}
\alias{Xdata-class}
\title{Xdata -- A class representing a data source}
\description{
Most methods of the class are abstract,
however the \code{show}, \code{print},
\code{queries} methods will usually not
need to be redefined.
}
\details{
The \code{query} method is defined for
character resource arguments. It is tried
to transform the argument to an object; if
that succeeds, \code{query} is called again.
Derived methods that are also interpreting
resource as character should first call
this method via \code{callNextMethod}.
}

