\name{Xsparql-class}
\alias{Xsparql-class}
\title{A class for querying SPARQL end points}
\description{
  This class allows to run SELECT statement on SPARQL
  endpoints. The resource parameter is interpreted as
  SPARQL statement.
}
\details{
  See \code{Dbpedia} for examples.
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

