\name{xsparql}
\alias{xsparql}
\title{Constructor for Xsparql}
\usage{
  xsparql(url, nspace = "")
}
\arguments{
  \item{url}{sparql end point}

  \item{nspace}{character vector with short name /
  namespace expansions}
}
\value{
  a xsparql object
}
\description{
  Constructor for Xsparql
}

