\docType{methods}
\name{as.character}
\alias{as.character}
\alias{as.character,DirectoryLocation-method}
\alias{as.character,Resource-method}
\title{The as.character for Xdata/Resource}
\description{
  The \code{as.character} method has been defined for
  various classes, such as the \code{Resource} class and
  the \code{DirectoryLocation} class.
}

