% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{as.character}
\alias{as.character}
\alias{as.character,DirectoryLocation-method}
\alias{as.character,FileTarget-method}
\alias{as.character,SftpLocation-method}
\alias{as.character,WebLocation-method}
\title{String representation of an Xdata object}
\usage{
\S4method{as.character}{DirectoryLocation}(x)

\S4method{as.character}{FileTarget}(x)

\S4method{as.character}{SftpLocation}(x)

\S4method{as.character}{WebLocation}(x)
}
\arguments{
\item{x}{an Xdata object}
}
\description{
The \code{as.character} method is extended to
various Xdata objects in order to give a more
detailled description of the object instance.
}
\details{
Inherited classes should override this method
if necessary.
}

