% Generated by roxygen2 (4.0.2): do not edit by hand
\name{city_coords}
\alias{city_coords}
\title{Longitude and Latitude for Cities}
\usage{
city_coords(country = "DE")
}
\arguments{
\item{country}{two-character country code, default 'DE'}
}
\value{
an object of class UrlData
}
\description{
The UrlData object provides a resource 'CityCoordinates' that
takes a \code{city} parameter and returns a two-valued vector with
latitude and longitude.
}
\details{
Based on Stackoverflow solution by Jochem Donkers.
}
\references{
\href{http://stackoverflow.com/questions/13905098/how-to-get-the-longitude-and-latitude-coordinates-from-a-city-name-and-country-i}{Stackoverflow}
}

