% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Mashup-class}
\alias{Mashup-class}
\alias{datamart}
\title{Combine resource into a mashup object}
\usage{
datamart(..., clss = "Mashup")
}
\arguments{
\item{...}{named arguments of Xdata objects}

\item{clss}{name of the class to create. Default Mashup, must be inherited from this class.}
}
\description{
The \code{Mashup} class administers a list of Xdata objects. This
can be data objects representing different data sources
such as internal data or web data. It can also be
calculated data resources, so-called resource functions
of class \code{ResFunc}.

Constructor for Mashup objects
}
\details{
In a way, this class can be viewed as a make-like tool
for data. The resource functions can declare dependencies.
When a resource is requested by the \code{query} method,
the \code{Mashup} class takes care of the build order.
}
\examples{
getSlots("Mashup")
}
\seealso{
\code{\link{datamart}}
}

