% Generated by roxygen2 (4.0.2): do not edit by hand
\name{expenditures}
\alias{expenditures}
\title{Elasticity of Private Households}
\usage{
expenditures()
}
\description{
This is a proof of concept for a use case of \code{datamart} to be used
for organising internal datasets as well as some calculations and graphics
on these datasets.
}
\details{
The example uses data from the Germany's Sample survey of income and expenditure (Einkommens- und Verbrauchsstichprobe, EVS).
}
\examples{
xp <- expenditures()
queries(xp)
query(xp, "categories")
query(xp, "elasticity", categ="05")
}
\references{
Statistisches Bundesamt: Wirtschaftsrechnungen. Einkommens- und Verbrauchsstichprobe. Einnahmen und Ausgaben privater Haushalte. Fachserie 15 Heft 4.
}
\seealso{
\code{\link{internalData}}, \code{\link{datamart}}
}

