% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{iwv_online}
\alias{iwv_online}
\title{website traffic from IWV}
\usage{
iwv_online()
}
\value{
UrlData object
}
\description{
website traffic as tracked by iwv online.
Use yyyymm for query() as resource.
This is an example for the \code{urldata} function.
}
\references{
\href{http://en.wikipedia.org/wiki/Informationsgemeinschaft_zur_Feststellung_der_Verbreitung_von_Werbetraegern}{Wikipedia}
}
\seealso{
\code{\link{urldata}}
}

