% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{meta}
\alias{meta}
\alias{meta,Blogger-method}
\alias{meta,CsvData-method}
\alias{meta,DirectoryLocation-method}
\alias{meta,InternalData-method}
\alias{meta,MemoryLocation2-method}
\alias{meta,Pastebin-method}
\alias{meta,SftpLocation-method}
\alias{meta,WebLocation-method}
\alias{meta,Xdata-method}
\alias{meta,Xsparql-method}
\title{Verbose list of resources}
\usage{
meta(self, ...)

\S4method{meta}{Xdata}(self, ...)

\S4method{meta}{Xsparql}(self)

\S4method{meta}{InternalData}(self, ...)

\S4method{meta}{DirectoryLocation}(self)

\S4method{meta}{MemoryLocation2}(self, ...)

\S4method{meta}{SftpLocation}(self)

\S4method{meta}{Pastebin}(self)

\S4method{meta}{WebLocation}(self)

\S4method{meta}{CsvData}(self, ...)

\S4method{meta}{Blogger}(self)
}
\arguments{
\item{self}{an Location object}

\item{...}{additional parameters}
}
\description{
The \code{meta} method returns a data.frame with meta information
entities available at the location. By default,
zero rows are returned.
}
\details{
Inherited classes should override this method
if necessary.
}

