% Generated by roxygen2 (4.0.2): do not edit by hand
\name{netConnectGermany}
\alias{netConnectGermany}
\title{Basic Price Information on natural Gas in Germany}
\usage{
netConnectGermany(clss = "UrlData")
}
\arguments{
\item{clss}{Class name for the object, default UrlData2}
}
\value{
UrlData2
}
\description{
This function exposes an interface to the Netconnect Germany website
and allows the download of price data on natural gas.
}
\references{
\href{http://datenservice.net-connect-germany.de/Dokumente/NCG_XML_Interface_V1.6_de.pdf}{NCG}
}

