% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{put}
\alias{put}
\alias{put,BlogPostTarget,Blogger-method}
\alias{put,BlogPostTarget,DirectoryLocation-method}
\alias{put,BlogPostTarget,SftpLocation-method}
\alias{put,CsvData,DirectoryLocation-method}
\alias{put,CsvData,SftpLocation-method}
\alias{put,FileTarget,DirectoryLocation-method}
\alias{put,FileTarget,SftpLocation-method}
\alias{put,GoogleOAuth2,character-method}
\alias{put,MdFigure,DirectoryLocation-method}
\alias{put,MdFigure,MemoryLocation-method}
\alias{put,MdReport,Location-method}
\alias{put,SweaveReport,DirectoryLocation-method}
\alias{put,SweaveReport,missing-method}
\alias{put,Target,MemoryLocation2-method}
\alias{put,Target,character-method}
\alias{put,character,SftpLocation-method}
\title{Put a target}
\usage{
put(target, where, ...)

\S4method{put}{Target,character}(target, where, ...)

\S4method{put}{Target,MemoryLocation2}(target, where, ...)

\S4method{put}{FileTarget,DirectoryLocation}(target, where, overwrite = TRUE)

\S4method{put}{FileTarget,SftpLocation}(target, where, ...)

\S4method{put}{character,SftpLocation}(target, where, ...)

\S4method{put}{BlogPostTarget,DirectoryLocation}(target, where, ...)

\S4method{put}{BlogPostTarget,SftpLocation}(target, where, ...)

\S4method{put}{MdFigure,DirectoryLocation}(target, where, ...)

\S4method{put}{MdFigure,MemoryLocation}(target, where)

\S4method{put}{MdReport,Location}(target, where, draft = TRUE,
  overwrite = TRUE, ...)

\S4method{put}{SweaveReport,DirectoryLocation}(target, where, verbose = TRUE,
  ...)

\S4method{put}{SweaveReport,missing}(target, where, verbose = TRUE, ...)

\S4method{put}{CsvData,DirectoryLocation}(target, where, verbose = TRUE, ...)

\S4method{put}{CsvData,SftpLocation}(target, where, verbose = TRUE, ...)

\S4method{put}{GoogleOAuth2,character}(target, where, ...)

\S4method{put}{BlogPostTarget,Blogger}(target, where, verbose = TRUE, ...)
}
\arguments{
\item{target}{an object of class Target or derived}

\item{where}{an object of class Location or derived}

\item{...}{additional parameters}

\item{overwrite}{parameter for FileTarget/MdReport -- overwrite existing files? Default TRUE.}

\item{draft}{Parameter to MdReport, is draft? Logical.}

\item{verbose}{print diagnostic messages}
}
\description{
This generic creates the target at the given location.
}

