% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{queries}
\alias{queries}
\alias{queries,CsvData-method}
\alias{queries,InternalData-method}
\alias{queries,Location-method}
\alias{queries,Mashup-method}
\alias{queries,MemoryLocation2-method}
\alias{queries,ResFunc-method}
\alias{queries,UrlData-method}
\alias{queries,Xdata-method}
\alias{queries,Xsparql-method}
\title{List resources}
\usage{
queries(self)

\S4method{queries}{Xdata}(self)

\S4method{queries}{Xsparql}(self)

\S4method{queries}{InternalData}(self)

\S4method{queries}{UrlData}(self)

\S4method{queries}{Mashup}(self)

\S4method{queries}{Location}(self)

\S4method{queries}{MemoryLocation2}(self)

\S4method{queries}{ResFunc}(self)

\S4method{queries}{CsvData}(self)
}
\arguments{
\item{self}{an Xdata object}
}
\description{
The \code{queries} method returns a character vector of all
defined resources for the given data object.
}
\details{
The default (XData) implementation inspects definitions of the
\code{query} method. Inherited classes should override this method
if necessary.
}

