\name{Resource-class}
\alias{Resource-class}
\title{S4 base class to represent resources}
\description{
  This class is used as based class when defining new
  resources for the \code{query} method. The
  \code{resource} function takes a character and defines a
  S4 class derived from Resource. Thus dispatching on the
  signature is possible, which is the main reason for using
  this class.
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

