\docType{methods}
\name{show}
\alias{show}
\alias{show,Blogger-method}
\alias{show,DirectoryLocation-method}
\alias{show,Target-method}
\alias{show,Xdata-method}
\title{Show method for Xdata}
\arguments{
  \item{object}{an Xdata object}

  \item{object}{an Target object}

  \item{object}{an Target object}

  \item{object}{an Target object}
}
\description{
  Show method for Xdata

  Show method for Directory

  Show method for Target

  Show method for Blogger
}

