% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{show}
\alias{show}
\alias{show,Blogger-method}
\alias{show,DirectoryLocation-method}
\alias{show,FileTarget-method}
\alias{show,Pastebin-method}
\alias{show,SftpLocation-method}
\alias{show,Target-method}
\alias{show,WebLocation-method}
\alias{show,Xdata-method}
\title{Show Method for Xdata classes}
\usage{
\S4method{show}{Xdata}(object)

\S4method{show}{DirectoryLocation}(object)

\S4method{show}{Target}(object)

\S4method{show}{FileTarget}(object)

\S4method{show}{SftpLocation}(object)

\S4method{show}{Pastebin}(object)

\S4method{show}{WebLocation}(object)

\S4method{show}{Blogger}(object)
}
\arguments{
\item{object}{Xdata object}
}
\description{
The \code{show} method for the Xdata class has been adapted to display the class name.
Some inherited classes such DirectoryLocation or Blogger override this default definition.
}

