\name{sourceforge}
\alias{sourceforge}
\title{SourceForge -- query stats for sourceforge projects}
\usage{
  sourceforge(proj, from = "2008-01-01", clss = "UrlData3")
}
\arguments{
  \item{proj}{name of the project}

  \item{from}{when did the project start? Default
  "2008-01-01".}

  \item{clss}{which clss to instantiate, default
  "UrlData3"}
}
\description{
  SourceForge -- query stats for sourceforge projects
}
\references{
  \url{http://sourceforge.net/p/forge/documentation/}
}

