% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sourceforge}
\alias{sourceforge}
\title{SourceForge -- query stats for sourceforge projects}
\usage{
sourceforge(proj, from = "2008-01-01", clss = "UrlData")
}
\arguments{
\item{proj}{name of the project}

\item{from}{when did the project start? Default "2008-01-01".}

\item{clss}{which clss to instantiate, default "UrlData"}
}
\description{
This exposes part of the API of the sourceforge.net website.
The function is a proof of concept for the \code{urldata} function.
}
\references{
\href{http://sourceforge.net/p/forge/documentation/}{SourceForge}
}
\seealso{
\code{\link{urldata}}
}

