% Generated by roxygen2 (4.0.2): do not edit by hand
\name{strencrypt}
\alias{strencrypt}
\title{Obfuscate string}
\usage{
strencrypt(message)
}
\arguments{
\item{message}{character or character vector}
}
\value{
character
}
\description{
a fancy method to make the string unreadable
use strdecypt to revert (not safe across machines!)
}

