% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{UrlData-class}
\alias{UrlData-class}
\title{UrlData -- unified access to WWW resources}
\description{
This class provides the infrastructure to
scrape the web with a Extract, Transform, Load (ETL)
approach.
}
\details{
In most cases, it is not necessary to subclass \code{UrlData}.
The slots can be set by the \code{urldata} function and allow
to customize each step of the process.
}
\examples{
getSlots("UrlData")
}
\seealso{
\code{\link{urldata}}
}

