% Generated by roxygen2 (4.0.2): do not edit by hand
\name{urldata}
\alias{urldata}
\title{Constructor for UrlData objects}
\usage{
urldata(resource, template, extract.fct = readLines,
  transform.fct = identity, clss = "UrlData", ...)
}
\arguments{
\item{resource}{the name of the resource. Required.}

\item{template}{a pattern for the url. Must contain \%s for substitution. Required.}

\item{extract.fct}{a function that takes an URI and returns the raw data. Default readLines.}

\item{transform.fct}{a function that takes the raw data and returns the cleaned/transformed data. Default identity.}

\item{clss}{name of the class to create. Default UrlData, must be inherited from this class.}

\item{...}{parameters for the query. Must be named arguments, values can be characters (for defaults),
                       NULL, or functions.}
}
\description{
This function creates a web resource. As this, it
allows to map an Web API into an R S4 class.
}

