% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{Xsparql-class}
\alias{Xsparql-class}
\title{A class for querying SPARQL end points}
\description{
This class allows to run SELECT statement on SPARQL endpoints.
The resource parameter is interpreted as SPARQL statement.
}
\examples{
getSlots("Xsparql")
}
\seealso{
\code{\link{xsparql}}, \code{\link{dbpedia}}, \code{\link{enipedia}}, \code{\link{openei}}
}

