\name{xsparql}
\alias{xsparql}
\title{data object for sparql}
\usage{
  xsparql(url, ns = "")
}
\arguments{
  \item{url}{sparql end point}

  \item{ns}{character vector with short name / namespace
  expansions}
}
\value{
  a xsparql object
}
\description{
  data object for sparql
}
\author{
  Karsten Weinert \email{k.weinert@gmx.net}
}

