% Generated by roxygen2 (4.0.2): do not edit by hand
\name{xsparql}
\alias{xsparql}
\title{Constructor for Xsparql}
\usage{
xsparql(resource, url, statement, nspace = "", clss = "Xsparql", ...)
}
\arguments{
\item{resource}{the resource the query represents}

\item{url}{sparql end point}

\item{statement}{the SPARQL statement the query stands for}

\item{nspace}{character vector with short name / namespace expansions}

\item{clss}{class to create, default Xsparql}

\item{...}{named parameters for resources}
}
\value{
a Xsparql object
}
\description{
The function \code{xsparql} constructs a Xsparql object.
}

